GH-37 Add optional custom label format
This allows package users to provide a custom label format for the labels collection. While #37 has a workaround to re-map the labels from the final Collection, this has unnecessary overhead. Furthermore, it’s hard to discern between Year-week (for week interval) and Year-Month (for month interval) labels without additional logic. In my use case, the interval is user-configurable, so I cannot assume a fixed label format from the Trend result.

Signed-off-by: Tobias Grasse <tg@glancr.de>